refactor: rename canvas/shared.ts to canvas/_shared.ts
The _ prefix excludes it from the route-coverage meta-test, matching
the convention that _-prefixed files are support utilities, not route
surface.
