+++ packages/mcp-server/src/server/routes/canvas.ts
@@ -7,7 +7,7 @@ import { createLiveDocRouter } from './canvas/live-doc.js'
 import { createMaintenanceRouter } from './canvas/maintenance.js'
 import { createCanvasMetadataRouter } from './canvas/metadata.js'
 import { createRestoreRouter } from './canvas/restore.js'
-import { setBroadcastFn } from './canvas/shared.js'
+import { setBroadcastFn } from './canvas/_shared.js'
 import { createThumbnailsRouter } from './canvas/thumbnails.js'
 import { createVersionsRouter } from './canvas/versions.js'
 import { createWorkspacesRouter } from './canvas/workspaces.js'
