+++ packages/mcp-server/src/server/routes/canvas/_shared.ts
