+++ packages/mcp-server/src/server/routes/canvas/versions.ts
@@ -8,7 +8,7 @@ import { isCorruptStoredDataError } from '../../store/corrupt-stored-data.js'
 import { getDoc } from '../../store/doc-cache.js'
 import type { OperatorInfo, VersionStore } from '../../store/version-store.js'
 import { validateSlug, validateWorkspaceId, validationErrorBody } from '../../validators.js'
-import { defaultHumanDisplayName, handleCorruptStoredData } from './shared.js'
+import { defaultHumanDisplayName, handleCorruptStoredData } from './_shared.js'
 
 export interface VersionsRouterOptions {
   versionStore: VersionStore
